Сделал функцию, которая печатает байткод в человекочитаемом формате Исправил багу с NOP и созданием типа int16. А работало всё потому что две ошибки удачно друг на друга наложились!

+++ natrix_object.h
@@ -13,7 +13,7 @@
 typedef struct _Object {
     unsigned int type;
     union {
-        int16_t vInt;
+        int32_t vInt;
         double vDouble;
         uint8_t vChar;
         char *vString;
